Move polling functions into separate file
